Clean up utils/auth.py and rename to cognite_auth.
